+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ Changelog
 =========
 
 
+- Fix bug in ``get`` where an iterable default was iterated over instead of being returned when an object path wasn't found. Thanks urbnjamesmi1_!
+
+
 v3.4.6 (2016-10-31)
 -------------------